Add support for LSM trees
closes #168

+++ src/support/scratch.c
@@ -249,7 +249,7 @@ __wt_buf_catfmt(WT_SESSION_IMPL *session, WT_ITEM *buf, const char *fmt, ...)
  */
 int
 __wt_scr_alloc_func(WT_SESSION_IMPL *session,
-    uint32_t size, WT_ITEM **scratchp
+    size_t size, WT_ITEM **scratchp
 #ifdef HAVE_DIAGNOSTIC
     , const char *file, int line
 #endif
